get test to work

+++ exo-go/test_helpers/tutorial_feature_context.go
@@ -1,61 +1,79 @@
 package testHelpers
 
 import (
+	"fmt"
 	"io/ioutil"
-  "net/http"
-	"os"
+	"net/http"
+	"net/url"
 	"path"
 	"strings"
-  "url"
 
 	"github.com/DATA-DOG/godog"
+	"github.com/DATA-DOG/godog/gherkin"
+	"github.com/Originate/exosphere/exo-go/src/util"
+	"github.com/jaytaylor/html2text"
 )
 
 // TutorialFeatureContext defines the festure context for the tutorial
 // nolint gocyclo
 func TutorialFeatureContext(s *godog.Suite) {
 
-	s.Step(`^I see "([^"]*)" in the terminal$`, func(expectedText string) {
-		process.WaitForTextWithTimeout(expectedText, 100
+	s.Step(`^I see "([^"]*)" in the terminal$`, func(expectedText string) error {
+		return process.WaitForTextWithTimeout(expectedText, 1000)
 	})
 
-	s.Step(`^I am in an empty folder$`, func() {
-		appDir = path.Join(os.TempDir())
+	s.Step(`^I am in an empty folder$`, func() error {
 		return nil
 	})
 
-	s.Step(`^I cd into "([^"]*)"$`, func(dir string) {
+	s.Step(`^I cd into "([^"]*)"$`, func(dir string) error {
 		appName = dir
-		appDir = dir
+		appDir = path.Join("tmp", appName)
+		fmt.Println(appDir)
 		return nil
 	})
 
-	s.Step(`^my application contains the template folder "([^"]*)"$`, func(templateDir string) {
-		templateName = strings.Split(templateDir, "/")[1]
-		checkoutServiceTemplate(appDir, templateName)
+	s.Step(`^my application contains the template folder "([^"]*)"$`, func(templateDir string) error {
+		templateName := strings.Split(templateDir, "/")[1]
+		return checkoutServiceTemplate(appDir, templateName)
 	})
 
 	s.Step(`^waiting until I see "([^"]*)" in the terminal$`, func(expectedText string) error {
-		return process.WaitForTextWithTimeout(expectedText, 5000)
+		return process.WaitForTextWithTimeout(expectedText, 60000)
 	})
 
 	s.Step(`^I stop all running processes$`, func() error {
-    return process.Kill()
+		if process != nil {
+			return process.Kill()
+		}
+		return nil
 	})
 
-	s.Step(`^the file "([^"]*)":$`, func(filePath, content string) error {
-		return ioutil.WriteFile(path.Join(appDir, filePath), []byte(content), 0777)
+	s.Step(`^the file "([^"]*)":$`, func(filePath string, expectedContent *gherkin.DocString) error {
+		return ioutil.WriteFile(path.Join(appDir, filePath), []byte(expectedContent.Content), 0777)
 	})
 
-	s.Step(`^http:\/\/localhost:(\d+) displays:$`, func(content string) error {
-    resp, err := http.Get("http://example.com/")
-    responseData,err := ioutil.ReadAll(response.Body)
-    return validateTextContains(string(responseData, content)
+	s.Step(`^http:\/\/localhost:(\d+) displays:$`, func(port string, expectedContent *gherkin.DocString) error {
+		resp, err := http.Get(fmt.Sprintf("http://localhost:%s", port))
+		if err != nil {
+			return err
+		}
+		defer resp.Body.Close()
+		body, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+		formattedText, err := html2text.FromString(string(body))
+		if err != nil {
+			return err
+		}
+		text := util.Strip(`\*+\n|\-+\n`, formattedText)
+		return validateTextContains(text, expectedContent.Content)
 	})
 
-	s.Step(`^adding a todo entry called "([^"]*)" via the web application$`, func(entry) error {
-    url := "localhost:3000/todos"
-    req, err := http.PostForm(url, url.Values{"text":entry})
+	s.Step(`^adding a todo entry called "([^"]*)" via the web application$`, func(entry string) error {
+		_, err := http.PostForm("http://localhost:3000/todos", url.Values{"text": {entry}})
+		return err
 	})
 
 }
